requirements.txt
# Core dependencies for WNBA scripts (scores, standings, schedule)
requests>=2.31.0
pytz>=2024.1
python-dotenv>=1.1.0

# Note: If you need to run the API server (wnba_api.py), 
# install additional dependencies with: pip install -r requirements-api.txt 